+++ apps/examples/blinkid-core-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example-blinkid-core-api
 
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [a7ace02]
+  - @microblink/blinkid-core@7.2.2
+
 ## 1.1.5
 
 ### Patch Changes