+++ apps/examples/blinkid-simple/CHANGELOG.md
@@ -1,5 +1,11 @@
 # example-blinkid-simple
 
+## 1.1.6
+
+### Patch Changes
+
+- @microblink/blinkid@7.2.2
+
 ## 1.1.5
 
 ### Patch Changes